LayoutTests/platform/iphone-simulator/accessibility/aria-pressed-state-expected.txt

button 1
button 2
This test thats aria-pressed changes the traits returned by buttons

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS btn1.iphoneTraits != btn2.iphoneTraits is true
PASS successfullyParsed is true

TEST COMPLETE
